Creates a minimalist hex color clock
Using the Shoes graphics toolkit and the green_shoes gem, I render a clock in HHMMSS format (e.g., 3:50:20 am is presented as 035020) that displays the current local time refreshed every second. The text is white while the background is the hex color that corresponds to the current six-digit hexadecimal shown (e.g., [#035020 is a dark shade of green](http://www.color-hex.com/color/035020)).